Moved NPClassScore to its own file
